Fix compile errors on gcc

+++ peer/src/base/CryptoHandler.hpp
@@ -35,7 +35,7 @@ class CryptoHandler {
 
   static string sign(const PrivateKey& privateKey, const string& s) {
     string retval(s.length() + crypto_sign_BYTES, '\0');
-    uint64_t signedLength;
+    unsigned long long signedLength;
     SODIUM_FAIL(crypto_sign((uint8_t*)&retval[0], &signedLength,
                             (const uint8_t*)&s[0], s.length(),
                             privateKey.data()));
@@ -45,7 +45,7 @@ class CryptoHandler {
 
   static optional<string> unsign(const PublicKey& publicKey, const string& s) {
     string retval(s.length(), '\0');
-    uint64_t unsignedLength;
+    unsigned long long unsignedLength;
     if (crypto_sign_open((uint8_t*)&retval[0], &unsignedLength,
                          (const uint8_t*)&s[0], s.length(), publicKey.data())) {
       return nullopt;
